attached_assets/requirements_1759128566474.txt
psutil
requests
tabulate
wmi
pywin32
python-nmap
psycopg2-binary
jsonschema